requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
av==13.1.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
cachetools==6.2.2
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cloup==3.0.8
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.124.2
fastjsonschema==2.21.2
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.58.4
fqdn==1.5.1
fsspec==2025.9.0
gast==0.6.0
glcontext==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.55.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.74.0
grpcio-status==1.71.2
gymnasium==1.2.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.10
imageio-ffmpeg==0.6.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
isosurfaces==0.1.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.7.4.5
keras==3.11.2
kiwisolver==1.4.8
lark==1.2.2
libclang==18.1.1
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.3
moderngl==5.12.0
moderngl-window==3.1.1
mpmath==1.3.0
msgpack==1.1.1
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.1.1
omegaconf==2.3.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pdfminer.six==20251230
pdfplumber==0.11.9
pettingzoo==1.25.0
pillow==11.2.1
platformdirs==4.4.0
polars==1.33.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
pyglet==2.1.8
pyglm==2.8.2
Pygments==2.19.2
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==5.3.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.0.2
ray==2.49.1
referencing==0.36.2
regex==2025.11.3
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.1
scipy==1.16.1
screeninfo==0.8.1
Send2Trash==1.8.3
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
skia-pathops==0.8.0.post2
sniffio==1.3.1
soupsieve==2.8
srt==3.5.3
stable_baselines3==2.7.0
stack-data==0.6.3
starlette==0.50.0
sumolib==1.24.0
svgelements==1.9.6
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.18.1
text-unidecode==1.3
tf_keras==2.20.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traci==1.24.0
traitlets==5.14.3
transformers==4.57.5
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.201
ultralytics-thop==2.0.17
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
